project/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
pydantic==2.7.4
pydantic-settings==2.3.3
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
httpx==0.27.0
geopy==2.4.1
stripe==8.10.0
